fix: 🐛 move and rfactor cms-youtube component
