--- src/components/bnb-cms-youtube/scaffolding/bnb-cms-youtube.hbs
@@ -1,6 +1,0 @@
-<bnb-cms-youtube heading="lerne uns kennen" youtubeinfo='{"youtubeId":"spYbjeokOBs", "thumbnail":{"src":"../../_assets/img/testbild_landscape.jpg", "title":"Testbild 16:9", "objectFit":"cover", "objectPosition":"center", "srcset":[{"src":"../../_assets/img/testbild_300.jpeg", "width":300},
-    {"src":"../../_assets/img/testbild_landscape_600.jpeg", "width":600},
-    {"src":"../../_assets/img/testbild_landscape_1024.jpeg", "width":1024},
-    {"src":"../../_assets/img/testbild_landscape_1440.jpeg", "width":1440},
-    {"src":"../../_assets/img/testbild_landscape_1920.jpeg", "width":1920}]}}'
-    data-resources="[{paths: ['components/bnb-cms-youtube/index.js']}]"></bnb-cms-youtube>
