--- src/components/bnb-cms-youtube/styles.scss
@@ -1,21 +1,0 @@
-@import "node_modules/@biotope/element/lib/host.mixin";
-@import '../../resources/scss/styles.scss';
-
-@include host('bnb-cms-youtube') {
-    display: block;
-    background-color: $purple;
-
-    .component-wrapper {
-        @include cmsComponents();
-    }
-
-    .youtube-wrapper {
-        // height: 100%;
-        // width: 100%;
-        width: 899px;
-        height: 506px;
-        position: relative;
-        margin: 0 auto;
-    }
-
-}
