+++ src/components/bnb-youtube/styles.scss
@@ -1,44 +1,85 @@
 @import 'node_modules/@biotope/element/lib/host.mixin';
 @import '../../resources/scss/styles';
 
-@include host('bnb-youtube') {
-	display: flex;
-	position: absolute;
-	justify-content: center;
-	align-items: center;
-	height: inherit;
-	width: inherit;
+$icon-size-small: 50px;
+$icon-size-medium: 80px;
+$icon-size-large: 100px;
+$icon-size-max: 125px;
 
-	iframe {
-		width: 100%;
-		height: 100%;
+$indentation-small: calc(50% - (#{$icon-size-small} / 2));
+$indentation-medium: calc(50% - (#{$icon-size-medium} / 2));
+$indentation-large: calc(50% - (#{$icon-size-large} / 2));
+$indentation-max: calc(50% - (#{$icon-size-max} / 2));
+
+@include host('bnb-youtube') {
+	display: block;
+
+	.youtube-wrapper {
+		height: 0;
 		overflow: hidden;
-		object-fit: cover;
+		padding-top: 56.25%;
+		position: relative;
+		display: flex;
+		margin: 20px 0;
+		justify-content: center;
+	}
+
+	iframe,
+	.thumbnail-wrapper {
+		overflow: hidden;
 		position: absolute;
 		top: 0;
 		left: 0;
-		z-index: -1;
+		height: 100%;
+		width: 100%;
+
+		img {
+			height: 100%;
+			width: 100%;
+			object-fit: cover;
+			object-position: center;
+			position: absolute;
+			top: 0;
+			left: 0;
+		}
+
+			bnb-icon {
+				position: absolute;
+				top: $indentation-small;
+				left: $indentation-small;
+				height: $icon-size-small;
+				width: $icon-size-small;
+				z-index: 2;
+				fill: $purple;
+
+				@include fromBreakpointM {
+					top: $indentation-medium;
+					left: $indentation-medium;
+					height: $icon-size-medium;
+					width: $icon-size-medium;
+				}
+
+				@include fromBreakpointL {
+					top: $indentation-large;
+					left: $indentation-large;
+					height: $icon-size-large;
+					width: $icon-size-large;
+				}
+
+				@include fromBreakpointMax {
+					top: $indentation-max;
+					left: $indentation-max;
+					height: $icon-size-max;
+					width: $icon-size-max;
+				}
+			}
 	}
 
-	bnb-image {
-		// z-index: 0;
-		// width: 100%;
-		// height: 100%;
-		// overflow: hidden;
-		// object-fit: cover;
+	.thumbnail-wrapper {
+		z-index: 1;
 	}
 
 	.hidden {
 		display: none;
-	}
-
-	bnb-icon {
-		position: absolute;
-		top: 50%;
-		left: 50%;
-		height: 125px;
-		width: 125px;
-		z-index: 2;
-		fill: $purple;
 	}
 }
