+++ src/components/bnb-youtube/template.ts
@@ -8,13 +8,15 @@ export const template = (
 	refs,
 ): HTMLFragment => {
 	return html`
-		${data.thumbnail
-			? html`
+		<div class="youtube-wrapper">
+			${data.thumbnail
+				? html`
 					${createStyle(`
         :host img {
           object-fit: ${data.thumbnail.objectFit};
           object-position: ${data.thumbnail.objectPosition};
-        }`)}
+		}`)}
+				<div class="thumbnail-wrapper">
 					<img
 						ref=${refs.thumbnail}
 						src="${data.thumbnail.src}"
@@ -22,20 +24,21 @@ export const template = (
 						sizes="${data.thumbnail.sizes}"
 						alt="${data.thumbnail.title}"
 					/>
-					<span
-					ref=${refs.button} onclick=${data.onButtonClick}>
-						<bnb-icon icon="play">
+					<bnb-icon
+					ref=${refs.button} onclick=${data.onButtonClick}
+						 icon="play">
 						</bnb-icon>
-					</span>
-			  `
-			: ''}
-		<iframe
-			ref=${refs.video}
-			src=${`https://www.youtube.com/embed/${data.youtubeId}?modestbranding=true&showinfo=false&rel=0`}
-			frameborder="0"
-			allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
-			allowfullscreen
-		></iframe>
+				</div>
+					`
+					  : ''}
+			<iframe
+				ref=${refs.video}
+				src=${`https://www.youtube.com/embed/${data.youtubeId}?modestbranding=true&showinfo=false&rel=0`}
+				frameborder="0"
+				allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
+				allowfullscreen
+			></iframe>
+		</div>
 		${createStyle(styles)}
 	`;
 };
