+++ src/components/cms/bnb-cms-youtube/scaffolding/bnb-cms-youtube.hbs
@@ -1,0 +1,6 @@
+<bnb-cms-youtube heading="lerne uns kennen" youtubeinfo='{"youtubeId":"spYbjeokOBs", "thumbnail":{"src":"../../_assets/img/stage.jpg", "title":"Testbild 16:9", "objectFit":"cover", "objectPosition":"center", "srcset":[{"src":"../../_assets/img/testbild_300.jpeg", "width":300},
+    {"src":"../../_assets/img/stage_600.jpeg", "width":600},
+    {"src":"../../_assets/img/stage_1024.jpeg", "width":1024},
+    {"src":"../../_assets/img/stage_1440.jpeg", "width":1440},
+    {"src":"../../_assets/img/stage_1920.jpeg", "width":1920}]}}'
+    data-resources="[{paths: ['components/cms/bnb-cms-youtube/index.js']}]"></bnb-cms-youtube>
