+++ src/components/cms/bnb-cms-youtube/styles.scss
@@ -1,0 +1,32 @@
+@import 'node_modules/@biotope/element/lib/host.mixin';
+@import '../../../resources/scss/styles.scss';
+
+@include host('bnb-cms-youtube') {
+	display: block;
+	background-color: $purple;
+	margin-top: -40px;
+
+	.component-wrapper {
+		@include cmsComponents();
+        padding-bottom: 50px;
+		margin-bottom: 0;
+		margin-top: 0;
+
+		bnb-headline {
+            padding-top: 40px;
+			padding-bottom: 16px;
+		}
+
+		@include fromBreakpointL {
+            padding-bottom: 100px;
+			margin-bottom: 0;
+			margin-top: 0;
+		}
+	}
+
+	.youtube-wrapper {
+		max-width: 899px;
+		position: relative;
+		margin: 0 auto;
+	}
+}
