+++ src/components/cms/bnb-cms-youtube/template.ts
