+++ src/pages/01cms-components.bnb-cms-youtube.hbs
@@ -11,5 +11,5 @@ description: A simple demo of BnbCmsYoutube
 --}}
 
 {{> layouts/layout.default
-	contentMain="components/bnb-cms-youtube/scaffolding/bnb-cms-youtube"
+	contentMain="components/cms/bnb-cms-youtube/scaffolding/bnb-cms-youtube"
 }}
